# docs/contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: goodmoneying M1 Operations API
  version: 0.1.0
  description: 업비트 수집 파이프라인 M1 운영 서버 API 계약.
servers:
  - url: http://localhost:8000
security: []
...
paths:
...
  /v1/dashboard/summary/stream:
    get:
      tags:
        - 대시보드(Dashboard)
      summary: 운영 상태 대시보드 요약 SSE 스트림
      operationId: streamDashboardSummary
      parameters:
        - name: once
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: 테스트와 진단용 단일 이벤트 전송 여부
      responses:
        "200":
          description: 대시보드 요약을 dashboard 이벤트로 전송하는 SSE 스트림
          content:
            text/event-stream:
              schema:
                type: string
                description: "event: dashboard 형식의 SSE 스트림. data 필드는 DashboardSummary JSON이다."
...
  /v1/dashboard/targets:
    get:
      tags:
        - 대시보드(Dashboard)
      summary: 대시보드 수집 대상 패널 조회
      operationId: getDashboardTargets
      parameters:
        - $ref: "#/components/parameters/DashboardLimit"
        - $ref: "#/components/parameters/DashboardOffset"
      responses:
        "200":
          description: 수집 대상 행 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardTargetsResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /v1/dashboard/coverage:
    get:
      tags:
        - 대시보드(Dashboard)
      summary: 대시보드 커버리지 패널 조회
      operationId: getDashboardCoverage
      parameters:
        - $ref: "#/components/parameters/DashboardLimit"
        - $ref: "#/components/parameters/DashboardOffset"
      responses:
        "200":
          description: 데이터 유형별 커버리지 상태 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardCoverageResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /v1/dashboard/realtime-heatmap:
    get:
      tags:
        - 대시보드(Dashboard)
      summary: 대시보드 실시간 수집 히트맵 패널 조회
      operationId: getDashboardRealtimeHeatmap
      parameters:
        - $ref: "#/components/parameters/DashboardLimit"
        - $ref: "#/components/parameters/DashboardOffset"
      responses:
        "200":
          description: 대상별 실시간 수집 히트맵 행 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardRealtimeHeatmapResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /v1/dashboard/missing-ranges:
    get:
      tags:
        - 대시보드(Dashboard)
      summary: 대시보드 결측 상위 패널 조회
      operationId: getDashboardMissingRanges
      parameters:
        - $ref: "#/components/parameters/DashboardLimit"
        - $ref: "#/components/parameters/DashboardOffset"
      responses:
        "200":
          description: 결측 구간 상위 대상 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMissingRangesResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /v1/collection-targets:
    put:
      tags:
        - 수집(Collection)
      summary: 활성 수집 대상 저장
      operationId: updateCollectionTargets
      security:
        - OperatorToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCollectionTargetsRequest"
      responses:
        "200":
          description: 저장된 활성 수집 대상
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionTargetsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /v1/collection-targets/{instrumentId}/coverage-segments:
    get:
      tags:
        - 수집(Collection)
      summary: 코인별 수집 구간 lazy 조회
      operationId: getCollectionCoverageSegments
      parameters:
        - $ref: "#/components/parameters/InstrumentId"
      responses:
        "200":
          description: 확장 row에서 사용하는 구간형 수집 상태 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionCoverageSegmentsResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /v1/market-list/stream:
    get:
      tags:
        - 시장(Market)
      summary: 관심종목 목록 SSE 스트림
      operationId: streamMarketList
      parameters:
        - name: once
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: 테스트와 진단용 단일 이벤트 전송 여부
      responses:
        "200":
          description: 관심종목 가격 정보를 marketList 이벤트로 전송하는 SSE 스트림
          content:
            text/event-stream:
              schema:
                type: string
                description: "event: marketList 형식의 SSE 스트림. data 필드는 MarketListResponse JSON이다."
  /v1/instruments/{instrumentId}:
    get:
      tags:
        - 상품(Instrument)
      summary: 거래 상품 상세 조회
      operationId: getInstrumentDetail
      parameters:
        - $ref: "#/components/parameters/InstrumentId"
      responses:
        "200":
          description: 코인 상세 화면 View Model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstrumentDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /v1/instruments/{instrumentId}/candles:
    get:
      tags:
        - 상품(Instrument)
      summary: 캔들 조회
      operationId: getCandles
      parameters:
        - $ref: "#/components/parameters/InstrumentId"
        - name: unit
          in: query
          required: true
          schema:
            type: string
            enum: [1m, 3m, 5m, 10m, 15m, 30m, 60m, 240m, 1d]
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: 원천 또는 파생 캔들 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandleSeriesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
  /v1/instruments/{instrumentId}/ticker-snapshots:
    get:
      tags:
        - 상품(Instrument)
      summary: 현재가 스냅샷 조회
      operationId: getTickerSnapshots
      parameters:
        - $ref: "#/components/parameters/InstrumentId"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: 현재가 스냅샷 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TickerSnapshotsResponse"
  /v1/instruments/{instrumentId}/orderbook-summaries:
    get:
      tags:
        - 상품(Instrument)
      summary: 호가 요약 조회
      operationId: getOrderbookSummaries
      parameters:
        - $ref: "#/components/parameters/InstrumentId"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: 호가 요약 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderbookSummariesResponse"
  /v1/collection-runs:
    get:
      tags:
        - 수집(Collection)
      summary: 수집 실행 조회
      operationId: getCollectionRuns
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: 최근 수집 실행 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionRunsResponse"
  /v1/backfill/plans:
    post:
      tags:
        - 백필(Backfill)
      summary: 백필 계획 생성
      operationId: createBackfillPlan
      security:
        - OperatorToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBackfillPlanRequest"
      responses:
        "200":
          description: 보수적 추정치가 포함된 백필 계획
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackfillPlan"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /v1/backfill/jobs:
    get:
      tags:
        - 백필(Backfill)
      summary: 백필 작업 목록 조회
      operationId: getBackfillJobs
      responses:
        "200":
          description: 백필 작업 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackfillJobsResponse"
    post:
      tags:
        - 백필(Backfill)
      summary: 백필 작업 시작
      operationId: startBackfillJob
      security:
        - OperatorToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBackfillJobRequest"
      responses:
        "201":
          description: 생성된 백필 작업
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackfillJob"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /v1/backfill/jobs/{jobId}/{action}:
    post:
      tags:
        - 백필(Backfill)
      summary: 백필 작업 제어
      operationId: controlBackfillJob
      security:
        - OperatorToken: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: action
          in: path
          required: true
          schema:
            type: string
            enum: [pause, stop, resume, safe-restart]
      responses:
        "200":
          description: 갱신된 백필 작업
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackfillJob"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /v1/backfill/jobs/{jobId}:
    delete:
      tags:
        - 백필(Backfill)
      summary: 백필 작업 삭제
      operationId: deleteBackfillJob
      security:
        - OperatorToken: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 삭제 완료
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    OperatorToken:
      type: apiKey
      in: header
      name: X-Operator-Token
  responses:
    BadRequest:
      description: 요청이 도메인 규칙을 만족하지 않음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 운영 토큰이 없거나 올바르지 않음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 요청한 리소스가 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: 요청 형식 검증 실패
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
    InstrumentId:
      name: instrumentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    DashboardLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 100
    DashboardOffset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
  schemas:
...
    DashboardTargetsResponse:
      description: 대시보드 수집 대상 패널의 페이지 응답.
      type: object
      required: [items, total, limit, offset, recommendedRefreshSeconds, refreshedAt]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CollectionDashboardTarget"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        recommendedRefreshSeconds:
          type: integer
        refreshedAt:
          type: string
          format: date-time
    DashboardCoverageResponse:
      description: 대시보드 커버리지 패널의 페이지 응답.
      type: object
      required: [items, total, limit, offset, recommendedRefreshSeconds, refreshedAt]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CoverageStatus"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        recommendedRefreshSeconds:
          type: integer
        refreshedAt:
          type: string
          format: date-time
...
    DashboardRealtimeHeatmapResponse:
      description: 대시보드 실시간 수집 히트맵 패널의 페이지 응답.
      type: object
      required: [items, total, limit, offset, recommendedRefreshSeconds, refreshedAt]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RealtimeCollectionHeatmapRow"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        recommendedRefreshSeconds:
          type: integer
        refreshedAt:
          type: string
          format: date-time
...
    DashboardMissingRangesResponse:
      description: 대시보드 결측 구간 상위 패널의 페이지 응답.
      type: object
      required: [items, total, limit, offset, recommendedRefreshSeconds, refreshedAt]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MissingRangeSummary"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        recommendedRefreshSeconds:
          type: integer
        refreshedAt:
          type: string
          format: date-time
...
    CollectionCoverageSegmentsResponse:
      description: 거래 상품 상세 확장 시 lazy 조회하는 커버리지 구간 목록 응답.
      type: object
      required: [instrumentId, items]
      properties:
        instrumentId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/CoverageSegment"
...
    UpdateCollectionTargetsRequest:
      description: 운영자가 활성 수집 대상으로 저장할 거래 상품 ID 목록 요청.
      type: object
      required: [instrumentIds]
      properties:
        instrumentIds:
          type: array
          maxItems: 50
          items:
            type: integer
            format: int64
        reason:
          type: string
    CollectionTargetsResponse:
      description: 저장된 활성 수집 대상 거래 상품 목록 응답.
      type: object
      required: [targets]
      properties:
        targets:
          type: array
          items:
            $ref: "#/components/schemas/Instrument"
...